requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
python-decouple
python-jose
aioredis
redis
groq
httpx
bcrypt
passlib



